rename tabSelector
